updated python namespace to bdgenomics.mango
