+++ mango-python/bdgenomics/mango/test/QC_test.py
@@ -16,3 +16,34 @@
 # limitations under the License.
 #
 
+import sys
+from bdgenomics.mango.QC import QC
+from bdgenomics.mango.test import SparkTestCase
+
+from bdgenomics.adam.adamContext import ADAMContext
+from bdgenomics.adam.rdd import CoverageRDD
+
+
+class QCTest(SparkTestCase):
+
+
+    def test_coverage_distribution(self):
+
+        # load file
+        ac = ADAMContext(self.sc)
+        testFile = self.resourceFile("small.sam")
+        # read alignments
+
+        reads = ac.loadAlignments(testFile)
+
+        # convert to coverage
+        coverage = reads.toCoverage()
+        # coverage = CoverageRDD(reads._jvmRdd.toCoverage(), self.sc) # TODO change once API is fixed
+
+        qc = QC()
+
+        cd = qc.CoverageDistribution(coverage, False)
+
+        assert(len(cd) == 1)
+        assert(cd.pop() == (1.0, 1500))
+
